requirements.txt
flask==3.0.3
dash==2.17.0
dash_bootstrap_components==1.6.0
matplotlib==3.10.5
numpy=2.3.2
pandas==2.2.3
pdfplumber==0.11.7
pydicom==3.0.1
PyQt6==6.9.1
PyYAML==6.0.2
scikit_image==0.25.2
scipy==1.16.1
waitress>=2.1.2
werkzeug==3.0.6